Move UploadStatus handling
